#73: Extracted routes and antd theme out of main.tsx
