+++ tsconfig.json
@@ -23,7 +23,8 @@
         "baseUrl": "./",
         "paths": {
             "@components/*": ["src/components/*"],
-            "@features/*": ["src/features/*"]
+            "@features/*": ["src/features/*"],
+            "@pages/*": ["src/pages/*"]
         },
 
         /* i18n */
